Decision trees and linear regressions written
They both work, but unfortunately the decision trees take a long time to
run so I'll be doing that tomorrow. The linear regressions each take
less than a second.

The decision trees are run through a GridSearchCV to find the best
parameters for max_depth and max_leaf_nodes, in order to ensure the
trees are small enough to actually be interpretable when made into
images using export_graphviz.

The linear regressions have no parameters, so have no need for
GridSearchCV.